\name{bonferroniQci}
\alias{bonferroniQci}
\title{Implementing Bonferroni Q-test confidence intervals for beta according to Campbell and Yogo (2006)}
\usage{
bonferroniQci(r, x, lags)
}
\description{
returns ci_beta_lower, ci_beta_upper, ci_rho_refined_lower, ci_rho_refined_upper, delta_hat, lags, df_gls, sd(e)/sd(u)
}
\examples{
bonferroniQci(r, x, lags)
}
