% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheus.R
\name{check_data}
\alias{check_data}
\title{Function}
\usage{
check_data(
  y,
  x,
  maxPowerSingleTerms,
  maxPowerCrossTerms,
  maxPowerCrossTermsTerms,
  maxNumberTerms,
  criterion,
  formula,
  distance,
  kf,
  testSet,
  rowsToTest,
  rowsNumber,
  includeBackward,
  decomposition,
  complexity,
  verbose,
  glmFamily,
  weights,
  asDummy
)
}
\description{
Function
}
