% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_prob_mat.R
\name{cnt_prob_mat}
\alias{cnt_prob_mat}
\alias{validate_cnt_prob_mat}
\title{Create a probability matrix with counts record}
\usage{
cnt_prob_mat(..., count = NULL)

validate_cnt_prob_mat(x)
}
\arguments{
\item{...}{Named \code{prob} vectors or vectors coercible to \code{prob}}

\item{count}{A positive integer vector indicating the number of cases.
Defaults to 1L recycled to the length of the probability vectors.}

\item{x}{A \code{cnt_prob_mat} for validation}
}
\value{
a \code{cnt_prob_mat} vector
}
\description{
Create a probability matrix with counts record
}
\examples{
p1 <- prob(c(0.4, 2 / 3, NA_real_, 1L))
p2 <- 1 - p1
cnt_prob_mat(pr_one = p1, pr_two = p2, count = c(10L, 20L, 5L, 15L))
cnt_prob_mat(pr_one = p1, pr_two = p2) # count defaults to 1L
}
