% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheus.R
\name{compute_criterion}
\alias{compute_criterion}
\title{Function}
\usage{
compute_criterion(
  criterion,
  y,
  x,
  distance,
  formula,
  rowsToTest,
  decomposition,
  glmFamily,
  weights
)
}
\description{
Function
}
