% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_oracle}
\alias{connect_oracle}
\title{Initialisation de la connexion à la base de données.}
\usage{
connect_oracle()
}
\value{
dbConnection Connexion à la base de données oracle
}
\description{
Initialisation de la connexion à la base de données.
}
