% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_table_from_query}
\alias{create_table_from_query}
\title{Création d'une table à partir d'une requête SQL.}
\usage{
create_table_from_query(conn = NULL, output_table_name = NULL, query = NULL)
}
\arguments{
\item{conn}{Connexion à la base de données}

\item{output_table_name}{Nom de la table de sortie}

\item{query}{Requête SQL}
}
\description{
Création d'une table à partir d'une requête SQL.
}
