% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheus.R
\name{estimate_backward}
\alias{estimate_backward}
\title{Function}
\usage{
estimate_backward(
  removable,
  y,
  chosenDesignMatrix,
  information,
  criterion,
  numFitB,
  distance,
  formula,
  folds,
  testSet,
  rowsToTest,
  rowsNumber,
  decomposition,
  glmFamily,
  weights
)
}
\description{
Function
}
