% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheus.R
\name{estimate_forward}
\alias{estimate_forward}
\title{Function}
\usage{
estimate_forward(
  candidate,
  x,
  y,
  chosenDesignMatrix,
  information,
  criterion,
  numFit,
  distance,
  formula,
  folds,
  testSet,
  rowsToTest,
  rowsNumber,
  decomposition,
  glmFamily,
  weights
)
}
\description{
Function
}
