% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hospital_consultations.R
\name{extract_hospital_consultations}
\alias{extract_hospital_consultations}
\title{Extraction des consultations externes à l'hôpital (MCO).}
\usage{
extract_hospital_consultations(
  start_date,
  end_date,
  spe_codes = NULL,
  prestation_codes = NULL,
  patient_ids = NULL,
  output_table_name = NULL,
  conn = NULL
)
}
\arguments{
\item{start_date}{Date La date de début de la période sur laquelle extraire les consultations.}

\item{end_date}{Date La date de fin de la période sur laquelle extraire les consultations.}

\item{spe_codes}{character vector Les codes spécialités des médecins effectuant les consultations à extraire. Si \code{spe_codes} n'est pas fourni, les consultations de tous les spécialités sont extraites.}

\item{prestation_codes}{character vector Les codes des prestations à extraire. Si \code{prestation_codes} n'est pas fourni, les consultations de tous les prestations sont extraites. Les codes des prestations sont disponibles sur la page \href{https://documentation-snds.health-data-hub.fr/snds/fiches/actes_consult_externes.html#exemple-de-requetes-pour-analyse}{actes et consultations externes de la documentation SNDS}.}

\item{patient_ids}{data.frame Un data.frame contenant les paires
d'identifiants des patients pour lesquels les consultations doivent être
extraites. Les colonnes de ce data.frame doivent être "BEN_IDT_ANO" et
"BEN_NIR_PSA" (en majuscules). Les "BEN_NIR_PSA" doivent être tous les
"BEN_NIR_PSA" associés aux "BEN_IDT_ANO" fournis. Si \code{patients_ids} n'est pas
fourni, les consultations de tous les patients sont extraites.}

\item{output_table_name}{character Le nom de la table de sortie dans la base de données. Si \code{output_table_name} n'est pas fourni, une table de sortie intermédiaire est créée.}

\item{conn}{dbConnection La connexion à la base de données. Si \code{conn} n'est pas fourni, une connexion à la base de données est initialisée.}
}
\value{
Un data.frame contenant les consultations. Les colonnes sont les suivantes :
\itemize{
\item BEN_IDT_ANO : Identifiant bénéficiaire anonymisé (seulement si patient_ids non nul)
\item NIR_ANO_17 : NIR anonymisé
\item EXE_SOI_DTD : Date de la délivrance
\item ACT_COD : Code de l'acte
\item EXE_SPE : Code de spécialité du professionnel de soin prescripteur
}
}
\description{
Cette fonction permet d'extraire les consultations à l'hôpital en MCO. Les consultations dont les dates
EXE_SOI_DTD sont comprises entre start_date et end_date sont extraites.
}
\details{
Si spe_codes est renseigné, seules les consultations des spécialités
correspondantes sont extraites.

Si prestation_codes est renseigné, seules les consultations des prestations correspondantes sont extraites.

Si patients_ids est fourni, seules les délivrances
de médicaments pour les patients dont les identifiants
sont dans patients_ids sont extraites.
}
\examples{
\dontrun{
extract_hospital_consultations(
  start_date = as.Date("2019-01-01"),
  end_date = as.Date("2019-12-31"),
  spe_codes = c("01", "02")
)
}
}
