% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_first_non_archived_year}
\alias{get_first_non_archived_year}
\title{Récupération de l'année non archivée la plus ancienne.}
\usage{
get_first_non_archived_year(conn)
}
\arguments{
\item{conn}{Connexion à la base de données}
}
\value{
Année non archivée la plus ancienne
}
\description{
Récupération de l'année non archivée la plus ancienne.
}
