% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheus.R
\name{pr}
\alias{pr}
\title{Function}
\usage{
pr(
  y,
  x,
  maxPowerSingleTerms,
  maxPowerCrossTerms = maxPowerSingleTerms,
  maxPowerCrossTermsTerms = maxPowerCrossTerms - 1,
  maxNumberTerms = NULL,
  criterion = "AIC",
  formula = NULL,
  distance = "RMSEP",
  kf = 2,
  testSet = "CV",
  rowsToTest = NULL,
  rowsNumber = NULL,
  includeBackward = FALSE,
  decomposition = "QR",
  complexity = 0,
  verbose = TRUE,
  glmFamily = gaussian(link = "identity"),
  weights = rep(1, length(y)),
  initialModel = NULL,
  asDummy = NULL
)
}
\description{
Function
}
