% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob}
\alias{prob}
\alias{validate_prob}
\alias{is_prob}
\alias{as_prob}
\title{Create a probability vector}
\usage{
prob(x = double())

validate_prob(x)

is_prob(x)

as_prob(x)
}
\arguments{
\item{x}{A double vector of probabilities}
}
\value{
a \code{prob} vector
}
\description{
A probability vector here is used as a vector where all entries are between
0 and 1 or are missing.
}
\examples{
prob(c(0.4, 2 / 3, NA_real_, 1L))
}
