% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_mat.R
\name{prob_mat}
\alias{prob_mat}
\title{Create a probability matrix record}
\usage{
prob_mat(...)
}
\arguments{
\item{...}{Named \code{prob} vectors or vectors coercible to \code{prob}}
}
\value{
a \code{prob} vector
}
\description{
A collection of \code{prob} vectors which sum to 1 rowwise.
}
\examples{
p1 <- prob(c(0.4, 2 / 3, NA_real_, 1L))
p2 <- 1 - p1
prob_mat(pr_one = p1, pr_two = p2)
}
