% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{prob_mat_from_matrix}
\alias{prob_mat_from_matrix}
\alias{as.matrix.prob_mat}
\title{Convert matrix into prob_mat}
\usage{
prob_mat_from_matrix(x, ...)

\method{as.matrix}{prob_mat}(x, ...)
}
\arguments{
\item{x}{A prob_mat or matrix}

\item{...}{Ignored.}
}
\value{
\code{prob_mat_from_matrix()} returns a \code{prob_mat} and \code{as.matrix()} returns a matrix
}
\description{
Convert matrix into prob_mat
}
\examples{
p1 <- prob(c(0.4, 2 / 3, NA_real_, 1L))
p2 <- 1 - p1
pm <- prob_mat(pr_one = p1, pr_two = p2)
m <- as.matrix(pm)
m
prob_mat_from_matrix(m)
}
