% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probable-package.R
\docType{package}
\name{probable-package}
\alias{probable}
\alias{probable-package}
\title{probable: Classes for Probabilities}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
