\name{sizeDistortionTest}
\alias{sizeDistortionTest}
\title{Implementing sizeDistortionTest according to Campbell and Yogo (2006)}
\usage{
sizeDistortionTest(r, x, lags)
}
\description{
returns TRUE if null hypothesis that the actual size exceeds 0.075 is rejected
}
\examples{
sizeDistortionTest(r, x, lags)
}
