% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prometheus.R
\name{update_candidate}
\alias{update_candidate}
\title{Function}
\usage{
update_candidate(
  candidate,
  chosen,
  chosenMatrix,
  maxPowerSingleTerms,
  maxPowerCrossTerms,
  maxPowerCrossTermsTerms
)
}
\description{
Function
}
