% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob-coerce.R
\name{vec_ptype2.prob}
\alias{vec_ptype2.prob}
\title{\code{prob} Coercion}
\usage{
\method{vec_ptype2}{prob}(x, y, ...)
}
\arguments{
\item{x, y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a vector of the same length, as class \code{prob} if convertible,
otherwise \code{double}
}
\description{
Coercion methods for \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}
}
\keyword{internal}
