\docType{package}
\name{pr-package}
\alias{pr}
\alias{pr-package}
\title{Support function for pupil dilation analysis}
\description{
  Support functions for analyzing pupil dilation data
}
\details{
  \tabular{ll}{ Package: \tab pr\cr Type: \tab Package\cr
  Version: \tab 0.1\cr Date: \tab 2008-08-24\cr License:
  \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }

  A set of functions used in the analysis of pupil dilation
  data, ranging from functions to detect blinks, do
  smoothing and filtering, and some functions for
  baselining and simple analyses.
}
\examples{
mean(13)
}
\author{
  Hedderik van Rijn \email{hedderik@van-rijn.org}, Jacolien
  van Rij
}
\references{
  \url{http://www.van-rijn.org/pR}
}
\keyword{package}

